+++ crawler/crawling/log_retry_middleware.py
@@ -61,7 +61,8 @@ class LogRetryMiddleware(object):
         if self.settings['STATS_STATUS_CODES']:
             self.redis_conn = redis.Redis(host=self.settings.get('REDIS_HOST'),
                                           port=self.settings.get('REDIS_PORT'),
-                                          db=settings.get('REDIS_DB'))
+                                          db=settings.get('REDIS_DB'),
+                                          decode_responses=True)
 
             try:
                 self.redis_conn.info()