+++ utils/tests/test_redis_queue.py
@@ -29,14 +29,14 @@ class TestBase(TestCase):
     def test_encode(self):
         q = Base(MagicMock(), 'key', pickle)
         # python pickling is different between versions
-        data = pickle.dumps('cool', protocol=-1)
+        data = pickle.dumps('cool', protocol=-1).decode('latin1')
         self.assertEquals(q._encode_item('cool'), data)
         q2 = Base(MagicMock(), 'key', ujson)
         self.assertEquals(q2._encode_item('cool2'), '"cool2"')
 
     def test_decode(self):
         q = Base(MagicMock(), 'key', pickle)
-        self.assertEquals(q._decode_item(b"\x80\x02U\x04coolq\x00."), 'cool')
+        self.assertEquals(q._decode_item(u"\x80\x02U\x04coolq\x00."), 'cool')
 
         q2 = Base(MagicMock(), 'key', ujson)
         self.assertEquals(q2._decode_item('"cool2"'), 'cool2')