+++ docs/topics/changelog.rst
@@ -10,7 +10,12 @@ Scrapy Cluster 1.3
 
 Date: ??/??/????
 
-- todo
+- Add Python 3 support
+
+Crawler
+^^^^^^^
+
+- Improved request to dictionary serialization support
 
 Scrapy Cluster 1.2
 ------------------